go.mod: roll back github.com/containerd/cgroups to v1.0.4 release
This was updated in d5e2962fb33c358519ed4e10f37fd37a7841fd2b, but we
only needed the ebpf update. As nothing depends on this module anymore,
other than for the stats package (which didn't change in between), we
can (for now) roll it back to v1.0.4, and just force the newer ebpf
package.

Things rolled back (doesn't affect vendored code);

https://github.com/containerd/cgroups/compare/7083cd60b721..v1.0.4

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

Containerd-test-commit: 9726cd5485f316cb744d9bd6a70151a46e4ea10f